server/pyproject.toml
[tool.poetry]
name = "ai-project"
version = "0.1.0"
description = ""
authors = ["Kamil Pawlik <kampawlik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
